main.js override ip in meta.links
The ip address is already provided in ip field.
The ip may be overriden from switches.json and it can be public.
But that ip inside of links is always private so it useless.
To keep backward compatibility we remove for links returned by old kankun-json their http://ip part and replace it with base_url.